+++ miniconf/src/lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(not(any(test, doctest, feature = "std")), no_std)]
-#![cfg_attr(feature = "json-core", doc = include_str!("../README.md"))]
+#![cfg_attr(feature = "json-core", doc = include_str!("../../README.md"))]
 #![cfg_attr(not(feature = "json-core"), doc = "Miniconf")]
 #![deny(rust_2018_compatibility)]
 #![deny(rust_2018_idioms)]
@@ -18,14 +18,6 @@ mod option;
 mod json_core;
 #[cfg(feature = "json-core")]
 pub use json_core::*;
-
-#[cfg(feature = "mqtt-client")]
-mod mqtt_client;
-#[cfg(feature = "mqtt-client")]
-pub use mqtt_client::*;
-
-#[cfg(feature = "mqtt-client")]
-pub use minimq;
 
 // re-export for proc-macro
 #[doc(hidden)]